+++ ui_automation_tests/features/case_notes.feature
@@ -8,8 +8,7 @@ Feature: I want to add a note to an application and view notes
   Scenario: Add a new valid case note
     Given I go to exporter homepage and choose Test Org
     And I create a standard application via api
-    When I click on applications
-    And I click on application previously created
+    When I go to application previously created
     And I click the notes tab
     And I enter "This is a note on my application!" for case note
     And I click cancel button
@@ -22,21 +21,10 @@ Feature: I want to add a note to an application and view notes
   Scenario: Add a case note with too many characters
     Given I go to exporter homepage and choose Test Org
     And I create a standard application via api
-    When I click on applications
-    And I click on application previously created
+    When I go to application previously created
     And I click the notes tab
     And I enter "the maximum limit" for case note
     Then case note warning is "You have 0 characters remaining"
     When I enter "the maximum limit plus 1" for case note
     Then case note warning is "You have 1 character too many"
     And post note is disabled
-
-  @LT_920_prohibit_adding_case_note_in_terminal_status @regression
-  Scenario: Unable to add case note
-    Given I go to exporter homepage and choose Test Org
-    And I create a standard application via api
-    When I click on applications
-    And my application has been withdrawn
-    And I click on application previously created
-    And I click the notes tab
-    Then the case note text area is not present
