+++ ui_automation_tests/features/edit_open_application.feature
@@ -8,8 +8,7 @@ Feature: I want to be able to edit and update an active application
   Scenario: Edit an open application
     Given I go to exporter homepage and choose Test Org
     And I create an open application via api
-    When I click on applications
-    And I click on application previously created
+    When I go to application previously created
     And I click edit application
     And I choose to make major edits
     And I click on open goods tile
