+++ ui_automation_tests/pages/goods_countries_page.py
@@ -8,12 +8,6 @@ class GoodsCountriesPage(BasePage):
     DESELECT_ALL_ID = "link-deselect-all"  # id
     SAVE_BUTTON = "button[type='submit']"
 
-    def select_all(self):
-        elements = self.driver.find_elements_by_css_selector(self.GOV_CHECKBOXES_INPUT)
-        for element in elements:
-            if not element.is_selected():
-                element.click()
-
     def select_all_link(self):
         self.driver.find_element_by_id(self.SELECT_ALL_ID).click()
