+++ ui_automation_tests/step_defs/test_withdraw_application.py
@@ -1,6 +1,7 @@
 from pytest_bdd import scenarios, then, when, parsers
 
 from pages.application_page import ApplicationPage
+from pages.submitted_applications_page import SubmittedApplicationsPages
 from shared import selectors
 from shared.functions import element_with_id_exists
 
@@ -31,4 +32,18 @@ def the_application_will_have_status(driver, status):
 def i_wont_be_able_to_see_the_withdraw_button(driver):
     driver.set_timeout_to(0)
     assert not element_with_id_exists(driver, ApplicationPage.BUTTON_WITHDRAW_APPLICATION_ID)
+    driver.set_timeout_to(10)
+
+
+@then("the edit application button is not present")
+def edit_button_not_present(driver):
+    driver.set_timeout_to(0)
+    assert len((ApplicationPage(driver).find_edit_application_button())) == 0
+    driver.set_timeout_to(10)
+
+
+@then("the case note text area is not present")
+def edit_button_not_present(driver):
+    driver.set_timeout_to(0)
+    assert len((SubmittedApplicationsPages(driver).find_case_note_text_area())) == 0
     driver.set_timeout_to(10)
